--- crates/lovely-mac/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "lovely-mac"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-name = "lovely"
-crate-type = ["cdylib"]
-
-[dependencies]
-lovely-core = { version ="0.7.1", path = "../lovely-core" }
-
-once_cell = "1.19.0"
-dobby-rs = "0.1.0"
-ctor = "0.2.7"
