--- crates/lovely-mac/run_lovely_macos.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-gamename="Balatro"
-defaultpath="/Users/$USER/Library/Application Support/Steam/steamapps/common/$gamename"
-
-export DYLD_INSERT_LIBRARIES=liblovely.dylib
-
-cd "$defaultpath"
-./$gamename.app/Contents/MacOS/love "$@"
