+++ crates/lovely-unix/Cargo.toml
@@ -11,4 +11,5 @@ crate-type = ["cdylib"]
 lovely-core = { version ="0.7.1", path = "../lovely-core" }
 
 once_cell = "1.19.0"
+dobby-rs = "0.1.0"
 ctor = "0.2.7"
